requirements.txt
matplotlib==3.4.2
pyserial==3.5
